Update observability.md
Hey Nikita, 

My name is Savan, and I manage Dev relationships at Middleware. 

I would like to request you to please accept this pull request to include our tool in the list. 

We are a full-stack observability platform & as a token of appreciation (if you need), we can give you a few free licenses to our tool. :) 

Hoping to get accepted, 

Savan